#235 [FE refactor]철자고침 콘솔지움
